Change the direction of the outline arrow
Before this change the arrow on the collapsed outline pointed right,
and the arrow on the expanded outline pointed down.
After this change the arrow on the collapsed outline points down,
and the arrow on the expanded outline points up.

The previous arrow makes sense only when the arrow is located
on the left from the outline heading, thus indicating that the
hidden content is on the right.

But when the arrow is located on the right from the outline heading,
then another logic makes more sense with the down arrow indicating
there is hidden contents below.